session reorg ok

+++ web/home/src/main/java/org/yaen/starter/web/home/interceptors/SessionInterceptor.java
@@ -8,14 +8,14 @@ import org.springframework.session.SessionRepository;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
 import org.yaen.starter.common.util.utils.StringUtil;
-import org.yaen.starter.web.home.contexts.CookieSessionContext;
-import org.yaen.starter.web.home.contexts.LocalSessionContext;
+import org.yaen.starter.web.home.sessions.CookieContext;
+import org.yaen.starter.web.home.sessions.SessionStorage;
 
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 
 /**
- * session interceptor, mostly for redis session
+ * session interceptor
  * 
  * @author Yaen 2016年5月19日下午3:33:09
  */
@@ -26,13 +26,13 @@ public class SessionInterceptor extends HandlerInterceptorAdapter {
 	@Setter
 	private int sessionMaxInactiveIntervalInSeconds = 108000;
 
-	/** session repository, maybe any type */
+	/** session client, save session id */
+	@Setter
+	private CookieContext cookieContext;
+
+	/** session repository, save session content */
 	@Setter
 	private SessionRepository<ExpiringSession> sessionRepository;
-
-	/** session context */
-	@Setter
-	private CookieSessionContext sessionContext;
 
 	/**
 	 * pre, init session
@@ -46,25 +46,35 @@ public class SessionInterceptor extends HandlerInterceptorAdapter {
 		// call super
 		if (super.preHandle(request, response, handler)) {
 
-			// get session id from cookie
-			String sessionId = this.sessionContext.getSessionId(request);
-
-			// get session from given repository
 			ExpiringSession session = null;
+
+			// get session id from cookie
+			String sessionId = this.cookieContext.getSessionId(request);
+			log.debug("get session context id:{}", sessionId);
+
+			// get session from repository
 			if (StringUtil.isNotEmpty(sessionId)) {
 				session = sessionRepository.getSession(sessionId);
+				log.debug("get session:{}", session);
+			}
+
+			// check expire
+			if (session != null && session.isExpired()) {
+				this.sessionRepository.delete(session.getId());
+				log.debug("session expired, deleted! sessionId:{}", session.getId());
+				session = null;
 			}
 
 			// create session if not exists
 			if (session == null) {
 				session = this.sessionRepository.createSession();
 				session.setMaxInactiveIntervalInSeconds(this.sessionMaxInactiveIntervalInSeconds);
-				this.sessionRepository.save(session);
-				log.debug("createSession success! sessionId:{}", session.getId());
+				log.debug("session created! sessionId:{}", session.getId());
 			}
 
-			// save session in thread local
-			LocalSessionContext.setSession(session);
+			// save session in local storage
+			SessionStorage.setLocalSession(session);
+			log.debug("save session in local storage");
 
 			return true;
 		}
@@ -81,18 +91,32 @@ public class SessionInterceptor extends HandlerInterceptorAdapter {
 	public void postHandle(HttpServletRequest request, HttpServletResponse response, Object handler,
 			ModelAndView modelAndView) throws Exception {
 		// get session from thread local
-		ExpiringSession session = LocalSessionContext.getSession();
+		ExpiringSession session = SessionStorage.getLocalSession();
+		log.debug("get session from local storage:{}", session);
 
-		// save or delete session, as may be changed by controller
-		if (session.isExpired()) {
-			this.sessionRepository.delete(session.getId());
-		} else {
-			session.setMaxInactiveIntervalInSeconds(this.sessionMaxInactiveIntervalInSeconds);
-			this.sessionRepository.save(session);
+		// deal session if any
+		if (session != null) {
+
+			// save or delete session, as may be changed by controller
+			if (session.isExpired()) {
+				// delete session
+				this.sessionRepository.delete(session.getId());
+
+				// delete cookie due to expired
+				this.cookieContext.setSessionId(response, session.getId(), true);
+
+				log.debug("session expired, deleted! sessionId:{}", session.getId());
+			} else {
+				// save session
+				session.setMaxInactiveIntervalInSeconds(this.sessionMaxInactiveIntervalInSeconds);
+				this.sessionRepository.save(session);
+
+				// save to cookie
+				this.cookieContext.setSessionId(response, session.getId(), false);
+
+				log.debug("session saved");
+			}
 		}
-
-		// set to cookie
-		this.sessionContext.setSessionId(response, session.getId(), session.isExpired());
 
 		// call super
 		super.postHandle(request, response, handler, modelAndView);
@@ -109,7 +133,7 @@ public class SessionInterceptor extends HandlerInterceptorAdapter {
 			throws Exception {
 
 		// delete local
-		LocalSessionContext.setSession(null);
+		SessionStorage.setLocalSession(null);
 
 		// call super
 		super.afterCompletion(request, response, handler, ex);
